CLParameters: remove unused properties and move to utils
