+++ pytissueoptics/rayscattering/opencl/utils/__init__.py
@@ -1,1 +1,2 @@
 from .CLKeyLog import CLKeyLog
+from .CLParameters import CLParameters
